requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pydantic==2.6.1
pandas==2.2.0
python-dotenv==1.0.1
sqlalchemy==2.0.27
